+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 #### Enhancements:
 
+- Add support to the core library API client to optionally raise a custom exception
+  class when a non-200 (and not 204) response is returned from an API request.
+- Raise `AnalyzerAPIError` when a non-200 response is returned from the API.
+
 
 
 #### Bug Fixes